+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/media/AudioMediaGroupContent.kt
@@ -1,0 +1,7 @@
+package dev.inmo.tgbotapi.types.message.content.media
+
+import dev.inmo.tgbotapi.types.InputMedia.AudioMediaGroupMemberInputMedia
+
+interface AudioMediaGroupContent : MediaGroupContent {
+    override fun toMediaGroupMemberInputMedia(): AudioMediaGroupMemberInputMedia
+}