+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/media/TextedMediaContent.kt
@@ -1,0 +1,6 @@
+package dev.inmo.tgbotapi.types.message.content.media
+
+import dev.inmo.tgbotapi.CommonAbstracts.TextedInput
+import dev.inmo.tgbotapi.types.message.content.abstracts.MediaContent
+
+sealed interface TextedMediaContent : MediaContent, TextedInput