+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/media/VoiceContent.kt
@@ -18,7 +18,7 @@ data class VoiceContent(
     override val media: VoiceFile,
     override val text: String? = null,
     override val textSources: TextSourcesList = emptyList()
-) : MediaContent, TextedInput {
+) : TextedMediaContent {
     override fun createResend(
         chatId: ChatIdentifier,
         disableNotification: Boolean,